Add response types file.
Add file response_types and move all response structures there.
